+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changelog for NeoFS Contract
 - New node format in the netmap contract with no limits to the map size (#445, #461, #462, #464)
 - Experimental object metadata events in the container contract (#448, #451, #456, #457)
 - Auto-cleaning for inactive nodes in the netmap contract (#460)
+- `getContainerData` and `getEACLData` methods to Container contract (#481)
 
 ### Changed
 - Minimal Go version to 1.22 (#353, #387)